backend/authentication/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-requests
python-dotenv
PyJWT
psycopg2-binary
Faker
channels[daphne]
channels_redis